Break out converters tests
